Funcionalidad autorizacion

+++ app/Http/Controllers/PermissionController.php
@@ -3,18 +3,27 @@
 namespace App\Http\Controllers;
 
 use App\Http\Controllers\Controller;
+use App\Http\Traits\HasAuthorization;
 use Illuminate\Http\Request;
 use Spatie\Permission\Models\Permission;
+use App\Models\User;
 
 class PermissionController extends Controller
 {
+    use HasAuthorization;
     /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
+    public function index(Request $request)
     {
+        if (!$this->isAuthorized("admin", $request->user())) {
+            return response()->json([
+                "message" => "User has not the right permissions."
+            ], 401);
+        }
+
         $obj_permissions = Permission::paginate(10);
 
         return $obj_permissions;
@@ -28,6 +37,12 @@ class PermissionController extends Controller
      */
     public function store(Request $request)
     {
+        if (!$this->isAuthorized("admin", $request->user())) {
+            return response()->json([
+                "message" => "User has not the right permissions."
+            ], 401);
+        }
+
         $validateFields = [
             "name" => "required",
             "guard_name" => "required"
@@ -56,6 +71,12 @@ class PermissionController extends Controller
      */
     public function show($id)
     {
+        if (!$this->isAuthorized("admin", User::class)) {
+            return response()->json([
+                "message" => "User has not the right permissions."
+            ], 401);
+        }
+
         $obj_permission = Permission::findOrFail($id);
 
         return $obj_permission;
@@ -70,6 +91,12 @@ class PermissionController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if (!$this->isAuthorized("admin", $request->user())) {
+            return response()->json([
+                "message" => "User has not the right permissions."
+            ], 401);
+        }
+
         $validateFields = [
             "name" => "required",
             "guard_name" => "required"
@@ -96,6 +123,11 @@ class PermissionController extends Controller
      */
     public function destroy($id)
     {
+        if (!$this->isAuthorized("admin", User::class)) {
+            return response()->json([
+                "message" => "User has not the right permissions."
+            ], 401);
+        }
         Permission::destroy($id);
 
         return response()->json(
